Changed class name and migration data type to work on linux

+++ database/migrations/2019_07_01_102959_create_posts_table.php
@@ -15,7 +15,7 @@ class CreatePostsTable extends Migration
     {
         Schema::create('posts', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('author_id');
+            $table->unsignedBigInteger('author_id');
             $table->string('cover_image');
             $table->string('alt_tag');
             $table->string('title');
